# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    Connect with your friends effortlessly using WASAText! Send and receive messages, 
    whether one-on-one or in groups, all from the convenience of your PC. 
    Enjoy seamless conversations with text or GIFs and easily stay in touch 
    through your private chats or group discussions.
  version: "1.0.0"
...
servers:
  - url: "http://localhost:3000"
    description: Development server
paths:
...
  /conversations/{conversationId}:
    parameters:
      - $ref: '#/components/parameters/conversationId'
    get:
      tags: ["Conversations"]
      summary: Get conversation messages
      description: Get all messages in a conversation in reverse chronological order
      operationId: getConversation
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Conversation messages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
              example:
                id: "conv123"
                type: "direct"
                participants:
                  - id: "user123"
                    name: "Maria"
                    photoUrl: "https://api.example.com/users/user123/photo"
                  - id: "user456"
                    name: "Marco"
                    photoUrl: "https://api.example.com/users/user456/photo"
                messages:
                  - id: "msg123"
                    senderId: "user123"
                    senderName: "Maria"
                    content: "Hello there!"
                    messageType: "text"
                    timestamp: "2023-12-07T10:30:00Z"
                    status: "read"
                    comments: []
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags: ["Conversations"]
      summary: Update conversation name
      description: Update the name of a group conversation. Only group conversations can be renamed.
      operationId: updateConversation
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for updating conversation name
              required:
                - name
              properties:
                name:
                  type: string
                  description: New name for the conversation/group
                  pattern: '^.+$'
                  minLength: 1
                  maxLength: 50
                  example: "Team Discussion"
            example:
              name: "Team Discussion"
      responses:
        '200':
          description: Conversation name updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
              example:
                id: "conv123"
                type: "group"
                name: "Updated Group Name"
                participants:
                  - id: "user123"
                    name: "Maria"
                    photoUrl: "https://api.example.com/users/user123/photo"
                  - id: "user456"
                    name: "Marco"
                    photoUrl: "https://api.example.com/users/user456/photo"
                messages: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: '#/components/parameters/conversationId'
    post:
      tags: ["Messages"]
      summary: Send a message
      description: |-
        Send a new message to a conversation. This endpoint handles:
        - Text messages: Set messageType to "text" and provide content
        - Photo messages: Use multipart/form-data with messageType "photo" and photo file
        
        Note: To forward existing messages, use the dedicated forwardMessage endpoint.
      operationId: sendMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Message content
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewMessage'
          multipart/form-data:
            schema:
              type: object
              description: Multipart form data for sending messages with optional photo attachments
              properties:
                messageType:
                  type: string
                  description: Type of message being sent
                  enum: ["text", "photo"]
                  example: "photo"
                content:
                  type: string
                  description: Text content (for text messages)
                  pattern: '^.+$'
                  minLength: 1
                  maxLength: 1000
                  example: "Hello there!"
                photo:
                  type: string
                  format: binary
                  description: Photo file binary data (for photo messages)
                  minLength: 1
                  maxLength: 10485760
                replyToMessageId:
                  type: string
                  description: ID of message being replied to (optional)
                  pattern: '^[a-zA-Z0-9]+$'
                  minLength: 1
                  maxLength: 50
                  example: "msg122"
              required:
                - messageType
      responses:
        '201':
          description: Message sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: "msg123"
                senderId: "user123"
                senderName: "Maria"
                content: "Hello there!"
                messageType: "text"
                timestamp: "2023-12-07T10:30:00Z"
                status: "sent"
                comments: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: ["Messages"]
      summary: Get conversation messages
      description: Get all messages in a conversation, separate from the conversation metadata
      operationId: getConversationMessages
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Messages retrieved successfully
          content:
            application/json:
              schema:
                type: object
                description: Response containing conversation messages
                properties:
                  messages:
                    type: array
                    description: List of messages in the conversation
                    minItems: 0
                    maxItems: 1000
                    items:
                      $ref: '#/components/schemas/Message'
                required:
                  - messages
              example:
                messages:
                  - id: "msg123"
                    senderId: "user123"
                    senderName: "Maria"
                    content: "Hello there!"
                    messageType: "text"
                    timestamp: "2023-12-07T10:30:00Z"
                    status: "read"
                    comments: []
                  - id: "msg124"
                    senderId: "user456"
                    senderName: "Marco"
                    content: "Hi Maria!"
                    messageType: "text"
                    timestamp: "2023-12-07T10:31:00Z"
                    status: "delivered"
                    comments: []
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: '#/components/parameters/conversationId'
      - $ref: '#/components/parameters/messageId'
    put:
      tags: ["Messages"]
      summary: Update message status
      description: Mark message as read or delivered
      operationId: updateMessageStatus
      security:
        - BearerAuth: []
      requestBody:
        description: New message status
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for updating message delivery status
              properties:
                status:
                  type: string
                  enum: ["delivered", "read"]
                  description: New delivery/read status for the message
                  example: "read"
              required:
                - status
      responses:
        '200':
          description: Message status updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 100
                    example: "Message status updated successfully"
              example:
                message: "Message status updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: ["Messages"]
      summary: Delete a message
      description: Delete a sent message
      operationId: deleteMessage
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Message deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/messages/{messageId}/photo:
    parameters:
      - $ref: '#/components/parameters/conversationId'
      - $ref: '#/components/parameters/messageId'
    get:
      tags: ["Messages"]
      summary: Get message photo
      description: Retrieve a photo from a message
      operationId: getMessagePhoto
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Photo retrieved successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
                description: PNG image file data
                minLength: 1
                maxLength: 10485760
            image/jpeg:
              schema:
                type: string
                format: binary
                description: JPEG image file data
                minLength: 1
                maxLength: 10485760
            image/gif:
              schema:
                type: string
                format: binary
                description: GIF image file data
                minLength: 1
                maxLength: 10485760
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Message not found or no photo available
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{conversationId}/forwarded_messages:
    parameters:
      - $ref: '#/components/parameters/conversationId'
    post:
      tags: ["Messages"]
      summary: Create a forwarded message
      description: Create a new message in the conversation by forwarding an existing message from another conversation
      operationId: forwardMessage
      security:
        - BearerAuth: []
      requestBody:
        description: Details of the message to forward
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for forwarding an existing message to this conversation
              properties:
                sourceConversationId:
                  type: string
                  description: ID of the conversation containing the original message
                  pattern: '^[a-zA-Z0-9]+$'
                  minLength: 1
                  maxLength: 50
                  example: "conv123"
                sourceMessageId:
                  type: string
                  description: ID of the original message to forward
                  pattern: '^[a-zA-Z0-9]+$'
                  minLength: 1
                  maxLength: 50
                  example: "msg456"
              required:
                - sourceConversationId
                - sourceMessageId
      responses:
        '201':
          description: Forwarded message created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
              example:
                id: "msg789"
                senderId: "user123"
                senderName: "Maria"
                content: "Original message content"
                messageType: "forward"
                timestamp: "2023-12-07T11:00:00Z"
                status: "sent"
                comments: []
                forwardFromMessageId: "msg456"
                forwardFromConversationId: "conv123"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    User:
      type: object
      description: User profile information
      properties:
        id:
          type: string
          description: Unique user identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "user123"
        name:
          type: string
          description: User's display name
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
          example: "Maria"
        photoUrl:
          type: string
          format: url
          description: URL to user's profile photo
          pattern: '^https?://.+$'
          minLength: 1
          maxLength: 500
          example: "https://api.example.com/users/user123/photo"
      required:
        - id
        - name
...
    MessageSummary:
      type: object
      description: Summary information about the last message in a conversation (subset of Message schema)
      properties:
        content:
          type: string
          description: Display content for the message (includes emoji prefixes for photos/forwards)
          pattern: '^.+$'
          minLength: 1
          maxLength: 1000
          example: "Hey, how are you doing?"
        messageType:
          type: string
          enum: ["text", "photo", "forward"]
          description: Type of message content
          example: "text"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the message was sent
          minLength: 1
          maxLength: 50
          example: "2023-12-07T10:30:00Z"
      required:
        - messageType
        - timestamp
    Conversation:
      type: object
      description: Full conversation details with messages
      properties:
        id:
          type: string
          description: Unique conversation identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "conv123"
        type:
          type: string
          enum: ["direct", "group"]
          description: Type of conversation
          example: "direct"
        participants:
          type: array
          description: List of conversation participants
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/User'
        messages:
          type: array
          description: List of messages in the conversation
          minItems: 0
          maxItems: 1000
          items:
            $ref: '#/components/schemas/Message'
        name:
          type: string
          description: Display name for the conversation (computed from participants or group name)
          pattern: '^.+$'
          minLength: 1
          maxLength: 50
          example: "Maria"
        photoUrl:
          type: string
          format: url
          description: URL to conversation photo (computed from participant or group photo)
          pattern: '^https?://.+$'
          minLength: 1
          maxLength: 500
          example: "https://api.example.com/users/user123/photo"
        lastMessage:
          $ref: '#/components/schemas/MessageSummary'
        unreadCount:
          type: integer
          minimum: 0
          maximum: 999999
          description: Number of unread messages in conversation
          example: 2
      required:
        - id
        - type
        - participants
        - messages
    Message:
      type: object
      description: Complete message information
      properties:
        id:
          type: string
          description: Unique message identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "msg123"
        senderId:
          type: string
          description: ID of the user who sent the message
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "user123"
        senderName:
          type: string
          description: Display name of the message sender
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "Maria"
        content:
          type: string
          description: Text content of the message (for text messages)
          pattern: '^.+$'
          minLength: 1
          maxLength: 1000
          example: "Hello there!"
        messageType:
          type: string
          enum: ["text", "photo", "forward"]
          description: Type of message content
          example: "text"
        photoUrl:
          type: string
          format: url
          description: URL to photo content (for photo messages)
          pattern: '^https?://.+$'
          minLength: 1
          maxLength: 500
          example: "https://api.example.com/conversations/conv123/messages/msg123/photo"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the message was sent
          minLength: 1
          maxLength: 50
          example: "2023-12-07T10:30:00Z"
        status:
          type: string
          enum: ["sent", "delivered", "read"]
          description: Delivery status of the message
          example: "read"
        comments:
          type: array
          description: List of emoji reactions/comments on the message
          minItems: 0
          maxItems: 50
          items:
            $ref: '#/components/schemas/Comment'
        replyToMessageId:
          type: string
          description: ID of the message this is replying to (if any)
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "msg122"
        forwardFromMessageId:
          type: string
          description: ID of the original message that was forwarded (for forward messages)
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "msg456"
        forwardFromConversationId:
          type: string
          description: ID of the original conversation the forwarded message came from (for forward messages)
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "conv789"
      required:
        - id
        - senderId
        - senderName
        - messageType
        - timestamp
        - status
        - comments
    NewMessage:
      type: object
      description: Data for creating a new message
      properties:
        content:
          type: string
          description: Text content of the message (required for text messages)
          pattern: '^.+$'
          minLength: 1
          maxLength: 1000
          example: "Hello there!"
        messageType:
          type: string
          enum: ["text", "photo"]
          description: Type of message to send
          example: "text"
        photo:
          type: string
          format: binary
          description: Photo file data (required for photo messages)
          minLength: 1
          maxLength: 10485760
        replyToMessageId:
          type: string
          description: ID of the message to reply to (optional)
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "msg122"
      required:
        - messageType
    Comment:
      type: object
      description: Emoji reaction/comment on a message
      properties:
        id:
          type: string
          description: Unique comment identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "comment123"
        userId:
          type: string
          description: ID of the user who made the comment
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "user456"
        userName:
          type: string
          description: Display name of the user who made the comment
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "John"
        emoticon:
          type: string
          description: Emoji character used as reaction
          pattern: '^.+$'
          minLength: 1
          maxLength: 10
          example: "😀"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the comment was made
          minLength: 1
          maxLength: 50
          example: "2023-12-07T10:35:00Z"
      required:
        - id
        - userId
        - userName
        - emoticon
        - timestamp
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g. missing fields, etc.)
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Unauthorized:
      description: The access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Forbidden:
      description: The user is not authorized to perform this operation
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |-
        Bearer authentication using user identifier.
        Use the identifier returned from the login endpoint.
  parameters:
...
    conversationId:
      name: conversationId
      in: path
      required: true
      description: Conversation identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "conv123"
    messageId:
      name: messageId
      in: path
      required: true
      description: Message identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "msg123"
...